Fixed mapview updating region when followLocation is false.
The mapview has two properties that define a region: (i) initialRegion
and (ii) region. Initially, we used the former to set an initial region
when opening the app. However, this disables tracking as long as the
followLocation state is true. That's sad, so we opted for region, but
when updating the position (the marker) this region was also re-set to
the mapView, instead of using the internal region. Since the event in
onPanDrag is inconclusive on what the new region is, I set it to null
like the professional retard that I am.